Some more cleanup

+++ test/jc_test.h
@@ -356,7 +356,7 @@ unsigned long long jc_test_get_time(void)
 {
     struct timeval tv;
     gettimeofday(&tv, NULL);
-    return tv.tv_sec * 1000000 + tv.tv_usec;
+    return (unsigned long long)tv.tv_sec * 1000000ULL + (unsigned long long)tv.tv_usec;
 }
 
 #endif
